split our ActiveRecord related methods
