+++ lib/active_interaction.rb
@@ -6,6 +6,7 @@ require 'active_interaction/version'
 require 'active_interaction/errors'
 
 require 'active_interaction/concerns/active_modelable'
+require 'active_interaction/concerns/active_recordable'
 require 'active_interaction/concerns/hashable'
 require 'active_interaction/concerns/missable'
 require 'active_interaction/concerns/transactable'
