+++ spec/active_interaction/base_spec.rb
@@ -313,31 +313,6 @@ describe ActiveInteraction::Base do
     end
   end
 
-  describe '#column_for_attribute(name)' do
-    let(:described_class) { InteractionWithFilter }
-    let(:column) { outcome.column_for_attribute(name) }
-
-    context 'name is not an input name' do
-      let(:name) { SecureRandom.hex }
-
-      it 'returns nil if the attribute cannot be found' do
-        expect(column).to be_nil
-      end
-    end
-
-    context 'name is an input name' do
-      let(:name) { InteractionWithFilter.filters.keys.first }
-
-      it 'returns a FilterColumn' do
-        expect(column).to be_a ActiveInteraction::FilterColumn
-      end
-
-      it 'returns a FilterColumn of type boolean' do
-        expect(column.type).to eql :float
-      end
-    end
-  end
-
   describe '#inputs' do
     let(:described_class) { InteractionWithFilter }
     let(:other_val) { SecureRandom.hex }
